pyproject.toml
[project]
name = "streamlit-syzygium-samarangense-classification"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "numpy>=2.0.2",
    "opencv-python>=4.10.0.84",
    "pillow>=11.0.0",
    "streamlit>=1.22.0",
    "tensorflow>=2.18.0",
]
